added migrations for openWeatherMap models

+++ JansenAndSixel/JansenAndSixel/Class/Main.cs
@@ -10,8 +10,8 @@ namespace JansenAndSixel.Class
         public double temp { get; set; }
         public int pressure { get; set; }
         public int humidity { get; set; }
-        public int temp_min { get; set; }
-        public int temp_max { get; set; }
+        public double temp_min { get; set; }
+        public double temp_max { get; set; }
 
     }
 }
