+++ JansenAndSixel/JansenAndSixel/Class/Wind.cs
@@ -8,7 +8,7 @@ namespace JansenAndSixel.Class
     public class Wind
     {
         public double speed { get; set; }
-        public int deg { get; set; }
+        public double deg { get; set; }
             
     }
 }
